+++ CHANGELOG.md
@@ -5,6 +5,9 @@ Note that since we don't clearly distinguish between a public and private interf
 
 
 ## [Unreleased]
+- Excluded common protein caps `NME` and `ACE` from the ligand selection query
+
+- Add screen-space shadow post-processing effect
 
 ## [v3.25.1] - 2022-11-20